pyproject.toml
[tool.poetry]
name = "themis"
version = "0.1.0"
description = ""
authors = ["Ádám Ruman <ruman.adam@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
networkx = "^2.6.3"
matplotlib = "^3.5.0"
frida = "^15.1.12"
frida-tools = "^10.4.1"
pyserde = "^0.5.3"
toml = "^0.10.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
